+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Next Release
 
+<br/>![New Features](/docs/assets/tags/new_features.svg)
+
+- [File or Project Open Dialog][xxxx]. <kbd>ctrl/cmd</kbd>+<kbd>o</kbd> will
+  display a new File or Project Open Dialog. Here you can open another project,
+  or pick a file from browser: the file will be inserted as a node to the
+  current graph.
+
 # Enso 2.0.0-alpha.8 (2021-06-09)
 
 <br/>![New Features](/docs/assets/tags/new_features.svg)
 
 #### Visual Environment
 
+<<<<<<< HEAD
 - [File browser component][1677]. A basic file browser component has been
   implemented, but is not integrated into the user interface yet.